+++ projects/tuist/fixtures/ios_app_with_static_frameworks/Modules/A/Sources/A.swift
@@ -1,5 +1,6 @@
 import B
 import C
+import PrebuiltStaticFramework
 
 public class A {
     public static let value: String = "aValue"
@@ -8,6 +9,7 @@ public class A {
         print("print from A")
         B.printFromB()
         C.printFromC()
+        print(StaticFrameworkClass().hello())
         print("---")
     }
 }